Moved the UITableViewDataSource methods to an extension.
